osgGeo: Fixed some C4018 warnings
Summary:
src\osgGeo\MarkerSet.cpp(291) : warning C4018: '>=' : signed/unsigned mismatch

src\osgGeo\TiledOffScreenRenderer.cpp(200) : warning C4018: '<' : signed/unsigned mismatch
src\osgGeo\TiledOffScreenRenderer.cpp(214) : warning C4018: '<' : signed/unsigned mismatch

src\osgGeo\TubeWellLog.cpp(286) : warning C4018: '<' : signed/unsigned mismatch

Test Plan: Still compiles on Linux. Tomorrow I'll check if the warnings are gone in CDash as well.

Reviewers: ding

Reviewed By: ding

Differential Revision: https://phabricator.dgbes.com:8443/D7266

+++ src/osgGeo/TiledOffScreenRenderer.cpp
@@ -28,7 +28,7 @@ $Id: TiledOffScreenRenderer.cpp 346 2014-02-07 09:31:17Z ding.zheng@dgbes.com $
 
 using namespace osgGeo;
 
-static unsigned char NOTRANSPARENCY = 255; 
+static unsigned char NOTRANSPARENCY = 255;
 
 static std::vector<std::pair<TiledOffScreenRenderer*,unsigned char> > _foregroundTransList;
 
@@ -171,7 +171,7 @@ bool TiledOffScreenRenderer::doRender()
 
 void TiledOffScreenRenderer::writeOutputFile(const std::string& fileName)
 {
-  if( !_imageCollector ) 
+  if( !_imageCollector )
 	return;
 
     osg::ref_ptr<osg::Image> outputImage = _imageCollector->getFinalImage();
@@ -182,7 +182,7 @@ void TiledOffScreenRenderer::writeOutputFile(const std::string& fileName)
 
 const osg::Image* TiledOffScreenRenderer::getOutput() const
 {
-    if( !_imageCollector  ) 
+    if ( !_imageCollector  )
 	return 0;
 
     return _imageCollector->getFinalImage();
@@ -197,28 +197,28 @@ void TiledOffScreenRenderer::setOutputBackgroundTransparency(unsigned char trans
 
 unsigned char TiledOffScreenRenderer::getForegroundTransparency() const
 {
-     for ( int idx=0; idx<_foregroundTransList.size(); idx++ )
-     {
-	    if ( _foregroundTransList[idx].first==this )
-	    {
+    for ( size_t idx=0; idx<_foregroundTransList.size(); idx++ )
+    {
+	if ( _foregroundTransList[idx].first==this )
+	{
 	    return _foregroundTransList[idx].second;
-	    }
-     }
+	}
+    }
 
- return 0;
+    return 0;
 }
 
 
 void TiledOffScreenRenderer::setForegroundTransparency(unsigned char transparency)
 {
-     for ( int idx=0; idx<_foregroundTransList.size(); idx++ )
-     {
-	    if ( _foregroundTransList[idx].first==this )
-	    {
+    for ( size_t idx=0; idx<_foregroundTransList.size(); idx++ )
+    {
+	if ( _foregroundTransList[idx].first==this )
+	{
 	    _foregroundTransList[idx].second = transparency;
 	    break;
-	    }
-     }
+	}
+    }
 }
 
 
@@ -311,7 +311,7 @@ void TiledOffScreenRenderer::OffscreenTileImageCollector::bindCameraToImage(
     camera->setProjectionMatrix(_currentProjectionMatrix * offsetMatrix);
 
     // Reattach cameras and new allocated images
-    camera->setRenderingCache( NULL );  
+    camera->setRenderingCache( NULL );
     camera->detach(osg::Camera::COLOR_BUFFER);
     camera->attach(osg::Camera::COLOR_BUFFER, image.get(), 0, 0);
 
@@ -323,7 +323,7 @@ void TiledOffScreenRenderer::OffscreenTileImageCollector::bindCameraToImage(
 	camera->detach(osg::Camera::DEPTH_BUFFER);
 	camera->attach(osg::Camera::DEPTH_BUFFER, depthImage.get(), 0, 0);
     }
-    
+
 }
 
 
@@ -348,7 +348,7 @@ void TiledOffScreenRenderer::OffscreenTileImageCollector::recordImages()
 		}
 
 		unsigned char* source = image->data(0, t);
-		unsigned char* target = _finalImage->data( 
+		unsigned char* target = _finalImage->data(
 		    col*(int)_tileSize.x(), t + row*(int)_tileSize.y());
 		memcpy(target, source, image->s() * 4 * sizeof(unsigned char));
 	    }
@@ -373,7 +373,7 @@ void TiledOffScreenRenderer::OffscreenTileImageCollector
 
     if ( !image || !depthImage )
 	return;
-    
+
     for ( int s=0; s<image->s();++s)
     {
 	const int imageOffset = (row * image->s() + s) * 4;
@@ -392,7 +392,7 @@ void TiledOffScreenRenderer::OffscreenTileImageCollector
 		imageData[3] = _foregroundTransparency;
 	    }
 	}
-	
+
     }
 }
